+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - Typed connect views [@fringer2423](https://github.com/fringer2423)
 - Typed QRCode views
+- Typed rate info views
 
 ### Изменено